+++ CHANGELOG.md
@@ -37,11 +37,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+## v0.14.0
+
 ### State Machine Breaking
 
+* [#224](https://github.com/babylonlabs-io/babylon/pull/224) Make injected checkpoint a standard tx
 * [#217](https://github.com/babylonlabs-io/babylon/pull/217) move voting power table to finality module
 * [#207](https://github.com/babylonlabs-io/babylon/pull/207) Rename total voting power
 to total bonded sat
+* [#204](https://github.com/babylonlabs-io/babylon/pull/204) Add babylon finality
+activation block height to start processing finality messages in `x/finality` params.
 * [#215](https://github.com/babylonlabs-io/babylon/pull/215) Implement ADR-29
 generalized unbonding handler